use pseudo element to uniformly apply rounded corners to table headers
since we cannot apply rounded corners to thead or tr - we have to apply them to
the table header cells but we only want them on the first (visible) and last
(visible) cells.

so we apply rounded corners to top edges of all table header cells and then we
add a pseudo element to thead to fill in the gaps in between the cells.