fix z-index containment within DataTable
data table applies z-index to create sticky headers/columns - but this can
cause these elevated rows/columns to bleed out of the table onto other elements
which have a lower z-index.

so we use isolation to create a new stacking context on the table to contain
the elevated elements within the table.