prevent rendering popover trigger on first render if using custom anchor
since radix internally uses context and mount lifecycle on PopoverAnchor to
determine if a custom anchor is being used and initially assumes no custom
anchor is being used.

as a result if a custom anchor is used the trigger will first render an anchor
and a button and then on the next render will remove the anchor and just render
the button. but this causes the first button ref to get swapped in DOM and
leads to inconsistent behavior if something targets that ref very quickly.