improve detection of addon click in input
by checking if clicking caused any child element to gain focus and only focus
the input otherwise. this way svg icons will still move focus to input even if
user somehow clicks exactly on a child path element.

and also detect clicks on input container to shift focus to input.

and in DateInput we prevent this and just open the calendar if user clicks
calendar icon

and in SearchInput, clear the value on pointerdown to prevent input from losing
focus