manually scroll selected item into view after mount
Downshift does not support rendering menu/listbox in portals and expects items
to be in DOM as soon as isOpen is true. previously we were delaying changing
highlightedIndex from -1 after menu was mounted to DOM but this causes a11y
issues where the menu is not properly announced by screen readers.

so now we change isOpen and highlightedIndex right away and manually scroll to
highlighted item on mount.