do not highlight selected item in Menu
Menu is for actions and not strictly for taking user input. so highlighting the
first selected option does not quite make sense - for example for an
action/dropdown menu where you have a mix of selectable options and action
options highlighting (and in turn scrolling to) the first selectable option
would not make sense.