easily allow selecting single date range
it might not be intuitive that selecting a single day range requires double
clicking the same date. so we make it easy to select a single day range by
storing the starting date and setting the range to that when closing the
popover if the user did not already select an end date.