optimize table rendering when switching to virtualization
render the previously used rows temporarily when toggling virtualization since
virtualizer returns an empty list for the first render and will cause all rows
to be re-mounted needlessly.

this is mostly an issue when using dynamic rows - like expandable rows - and
going from few rows to a large number of rows.